% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmPlot_function.R
\name{cmPlot}
\alias{cmPlot}
\title{Visualize a Confusion Matrix (ggplot)}
\usage{
cmPlot(cm, red_or_white, pred_first = TRUE, title, title.size = 20,
  axis.title.size = 15, axis.text.size = 18)
}
\arguments{
\item{cm}{a 3 column dataframe with columns for prediction and reference categroies and corresponding frequency counts (third column). 0 counts should be converted to NA}

\item{red_or_white}{character string c("red", "white") to specify plot coloring}

\item{pred_first}{logical, whether first column contains the prediction category (default is TRUE)}

\item{title}{character string to be used as plot title}

\item{title.size}{an integer controlling the size of the plot title. Defaults to 20.}

\item{axis.title.size}{an integer controlling the size of axis titles. Defaults to 15.}

\item{axis.text.size}{an integer controlling the size of axis text. Defaults to 18.}
}
\value{
a ggplot object visualizing the confusion matrix of reference versus predicted
}
\description{
Given a dataframe and a formula object, will return the necessary dependent
variable matrix and the model matrix needed to run the optimization for
multinomial regression for a given reference level.
}
