% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_likelihood.R
\name{data_to_mulitnomial}
\alias{data_to_mulitnomial}
\title{Reformats dataframe to multinomial regression form}
\usage{
data_to_mulitnomial(data, formula, ref_level = "8", intercept = T)
}
\arguments{
\item{data}{a dataframe}

\item{formula}{a formula object describing the regression to be run}

\item{ref_level}{a string which specifies the reference level of the factor dependent variable}

\item{intercept}{logical, whether to include an intercept or not in the regression (default is TRUE)}
}
\value{
a list containing:
\itemize{
\item An (n x (k-1)) response matrix of 0's and 1's
\item An (n x p) model matrix where n is the number of observations, p is the number of coefficients in the regression formula, and k is the number of possible levels of the dependent variable y.
}
}
\description{
Given a dataframe and a formula object, will return the necessary dependent
variable matrix and the model matrix needed to run the optimization for
multinomial regression for a given reference level.
}
