% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_likelihood.R
\name{fit_multinomial_regression}
\alias{fit_multinomial_regression}
\title{Fits multinomial regression through `optimx`}
\usage{
fit_multinomial_regression(data, formula, ref_level, intercept = T,
  method = "BFGS", trace = 0, tol = 1e-04, itters = 200)
}
\arguments{
\item{data}{a dataframe on which one would like to apply multinomial regression}

\item{formula}{a formula obect - in the style of `lm` which fits mulitnomial regression}

\item{ref_level}{a character, the reference level of the multinomial model}

\item{intercept}{default is TRUE, whether or not to include an intercept in the model}

\item{method}{default is BFGS, argument is passed to optimx}

\item{trace}{default is 0, passed to optimx}

\item{tol}{default is 10e-5, passed to optimx}

\item{itters}{default is 200, passed to optimx}
}
\value{
the fitted estimates of the beta coefficients in multinomial regression
}
\description{
Given a dataframe and a formula object, `fit_multinomial_regression` will run a
user-specified algorithm to fit multinomial linear regression for a user specified
reference level.
}
\details{
I should probably add the details of the assumptions of the model here.
}
