% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.multinomial.R
\name{generate.multinomial}
\alias{generate.multinomial}
\title{A function to create an n x j matrix of outcomes where the rows are subjects and columns are the ordinal levels of an ordinal outcome variable.}
\usage{
generate.multinomial(y, y.levels)
}
\arguments{
\item{y}{A vector of containing the values of an ordinal outcome variable.}

\item{y.levels}{A vector of the unique, ordinal levels of y.}
}
\value{
A matrix where the rows correspond to subjects and the columns correspond to levels of the outcome.
}
\description{
This function takes a numeric vector corresponding to observed ordinal values of a random variable, and a second vector specifying the ordered levels of y. This is not intended to be used independently.
}
