% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instructions.R
\name{get_instructions}
\alias{get_instructions}
\title{Extract instruction prompt}
\usage{
get_instructions(collapse = TRUE)
}
\arguments{
\item{collapse}{A logical: should the instructions be returned as a single
character string (the default) or a character vector.}
}
\value{
A character vector.
}
\description{
Pull the instruction prompt used by the package into R.
}
\examples{
substring(get_instructions(), 1, 23)
}
