% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.mod.R
\name{linear.mod}
\alias{linear.mod}
\title{A function to run linear regression}
\usage{
linear.mod(dat, intercept = TRUE)
}
\arguments{
\item{dat}{n*(p+1) data frame with the covariates and the response vector}

\item{intercept}{logical. whether to include the intercept or not. default is TRUE}
}
\value{
A list of 
\item{betahat}{A vector of beta estimates from the linear regression model.}
\item{yhat}{A vector of estimated response using the beta estimates from the model.}
}
\description{
This function runs linear regression model treating the response as continuous
}
\examples{

data(red_train)
rbeta = c('volatile.acidity', 'total.sulfur.dioxide','pH','alcohol','sulphates')
dat = red_train[,which(names(red_train)\%in\% c(rbeta,'quality'))]
linfit = linear.mod(dat, intercept=TRUE)
beta = linfit$betahat

}
