% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_likelihood.R
\name{multinomial_log_likelihood}
\alias{multinomial_log_likelihood}
\title{Computes the log likelihood for multinomial regression}
\usage{
multinomial_log_likelihood(y, X, b_vector)
}
\arguments{
\item{y}{an (n x (k-1)) matrix for the dependent variable as created by the data_to_multinomial function, where k is the number of categories for y.}

\item{X}{an (n x p) matrix where p is the number of predictors}

\item{b_vector}{a (p*(k-1)) vector}
}
\value{
the value of the mulitnomial regression log-likelihood
}
\description{
Given a matrix of responses, a design matrix, and a vector of beta coefficients
(must be a vector to be compatible with optim), it will return the value of the
log likelihood for multinomial regression.
}
\details{
I should probably add the details of the assumptions of the model here.
}
