% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.prop.odds.ll.R
\name{partial.prop.odds.ll}
\alias{partial.prop.odds.ll}
\title{A function to return the log likelihood for a partial proportional odds model}
\usage{
partial.prop.odds.ll(y, y.levels, in.data, int.vector,
  x.prop.odds = NULL, x.non.prop.odds = NULL, beta.prop.odds = NULL,
  beta.non.prop.odds = NULL)
}
\arguments{
\item{y}{A vector of containing the values of an ordinal outcome variable.}

\item{y.levels}{A vector of the unique, ordinal levels of y.}

\item{in.data}{The input data object of type data frame or matrix.}

\item{int.vector}{A vector of intercept estimates (one for each level of the outcome variable, except the top level). These need to be in the order of the levels of the outcome variable (from low to high).}

\item{x.prop.odds}{A design martrix (no intercept) for the variables assumed to have proportional odds.}

\item{x.non.prop.odds}{A design martrix (no intercept) for the variables assumed to not have proportional odds.}

\item{beta.prop.odds}{A vector of beta values for each predictor assumed to have proportional odds.}

\item{beta.non.prop.odds}{A matrix of beta values for each predictor assumed to not have proportional odds, where columns are the first j-1 levels of the ordinal outcome, and rows are betas.}
}
\value{
The log-likelihood for the model.
}
\description{
This function returns the og likelihood for a partial proportional odds model. This function should not be used independently.
}
