% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.partial.prop.odds.R
\name{predict.partial.prop.odds}
\alias{predict.partial.prop.odds}
\title{A function to predict probabilities for each level of an ordinal outcome for testing data using partial proportional odds model fit on training data.}
\usage{
\method{predict}{partial.prop.odds}(object, newdata, ...)
}
\arguments{
\item{object}{The output object from partial.prop.odds.mod run on a training dataset.}

\item{newdata}{A test dataset to be used for predictions. The outcome for test data must have the exact same levels as the training dataset}

\item{...}{Other parameters}
}
\value{
A data frame of predicted probabilities for each level of the ordinal outcome for each subject, where the subjects are rows and the levels of the outcome are columns. The final column is the most likely category.
}
\description{
This function takes a testing dataset and the output object from a partial proportional odds model fit on a training dataset, and predicts probabilities of each level of the outcome for each individual in the test data.
}
\examples{

data(red_test)
data(red_train)
starts <- coef(lm(quality ~ alcohol+ pH + volatile.acidity, data = red_train))
training.result <- partial.prop.odds.mod(y ="quality", in.data = red_train,
prop.odds.formula = ~ alcohol + pH,
beta.prop.odds.start = starts[2:3],
non.prop.odds.formula = ~ volatile.acidity,
beta.non.prop.odds.start = matrix(rep(starts[4], 5), nrow = 1),
method = "BFGS",
seed = c(14, 15), itnmax = 1000)
predictions <- predict(training.result, red_test)

}
