% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_likelihood.R
\name{predict_multinomial}
\alias{predict_multinomial}
\title{A function to do prediction with multinomial regression models}
\usage{
predict_multinomial(multi_fit, newdata)
}
\description{
A function to do prediction with multinomial regression models
}
