% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_chat.R
\name{recommend}
\alias{recommend}
\title{Invoke a chatbot to help write tidymodels code}
\usage{
recommend()
}
\value{
An object of class \code{Chat} (and \code{R6}) with the conversation.
}
\description{
This function uses a large language model (LLM) to process a user's questions
and return instructions and code.
}
\details{
The function specifically uses Antropic's Claude tools. You will need an API
key for that system. To set it, create an environmental variable named
\code{ANTHROPIC_API_KEY} or declare it in your local projects \code{.Renviron} file.
We suggest the latter using \code{\link[usethis:edit]{usethis::edit_r_environ()}}.
}
