% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_chat.R
\name{tm_chat}
\alias{tm_chat}
\title{Invoke a chatbot to help write tidymodels code}
\usage{
tm_chat(model = "gpt-4o-mini")
}
\arguments{
\item{model}{A character string for the LLM model.}
}
\value{
Nothing since users have to quit the discussion. However, the history
of the chat is saved in \code{.Last.chat.result}.
}
\description{
This function uses a large language model (LLM) to process a user's questions
and return instructions and code.
}
